+++ test/type_inference/common.lua
@@ -3609,7 +3609,7 @@ TEST '-2|-3|1' [[
 local <?n?>
 ]]
 
-TEST 'table' [[
+TEST 'enum A' [[
 ---@enum A
 local m = {}